fix review comments 29/07/20
