+++ app/controllers/foreman_inventory_upload/tasks_controller.rb
@@ -1,4 +1,4 @@
-module InventorySync
+module ForemanInventoryUpload
   class TasksController < ::ApplicationController
     def create
       selected_org = Organization.current
