--- webpack/ForemanInventoryUpload/Components/PageHeader/components/SyncButton/SyncButtonHelpers.js
@@ -1,3 +1,0 @@
-import { foremanUrl } from '../../../../../ForemanRhCloudHelpers';
-
-export const inventorySyncUrl = path => foremanUrl(`/inventory_sync/${path}`);
